fix(BadgeController): instantiate new TestContext in the constructor of the test class instead of constructor injection for DELETE api endpoint tests
