+++ ferrilata-devilline.IntegrationTests/Scenarios/ApiControllers/Badges/ApiDeleteBadgeTests.cs
@@ -1,28 +1,46 @@
+using System.Collections.Generic;
 using System.Linq;
 using System.Net;
 using System.Net.Http;
 using System.Threading.Tasks;
 using AngleSharp.Common;
 using ferrilata_devilline.IntegrationTests.Fixtures;
+using ferrilata_devilline.Services.Interfaces;
+using Newtonsoft.Json;
 using Xunit;
 
-namespace ferrilata_devilline.IntegrationTests
+namespace ferrilata_devilline.IntegrationTests.Scenarios.ApiControllers.Badges
 {
     [Collection("BaseCollection")]
     public class ApiDeleteBadgeTests
     {
         private readonly TestContext testContext;
+        private readonly ITokenService _tokenService;
+        private readonly string email;
 
-        public ApiDeleteBadgeTests(TestContext testContext)
+        
+        public ApiDeleteBadgeTests()
         {
-            this.testContext = testContext;
+            testContext = new TestContext();
+            _tokenService = testContext.TokenService;
+            email = "useremail@ferillata.com";
+        }
+
+        [Fact]
+        public async Task DeleteBadgeApi_IncorrectAuthentication_ShouldMessageEqual()
+        {
+            var request = new HttpRequestMessage(HttpMethod.Delete, "/api/badges/2");
+            var response = await testContext.Client.SendAsync(request);
+            var responseString = await response.Content.ReadAsStringAsync();
+            Assert.Equal("Unauthorized",
+                JsonConvert.DeserializeObject<Dictionary<string, string>>(responseString)["error"]);
         }
 
         [Fact]
         public async Task DeleteBadgeApi_CorrectAuthentication_ShouldReturnOK()
         {
             var request = new HttpRequestMessage(HttpMethod.Delete, "/api/badges/1");
-            request.Headers.Add("Authorization", "test");
+            request.Headers.Add("Authorization", "Bearer " + _tokenService.GenerateToken(email, true));
             var response = await testContext.Client.SendAsync(request, HttpCompletionOption.ResponseHeadersRead);
             Assert.Equal(HttpStatusCode.OK, response.StatusCode);
         }
@@ -31,7 +49,7 @@ namespace ferrilata_devilline.IntegrationTests
         public async Task DeleteBadgeApi_CorrectAuthentication_ShouldDeleteBadge()
         {
             var request = new HttpRequestMessage(HttpMethod.Delete, "/api/badges/1");
-            request.Headers.Add("Authorization", "test");
+            request.Headers.Add("Authorization", "Bearer " + _tokenService.GenerateToken(email, true));
             await testContext.Client.SendAsync(request, HttpCompletionOption.ResponseHeadersRead);
             Assert.True(testContext.Context.Badges.Count(badge => badge.BadgeId == 1) == 0);
         }
@@ -40,10 +58,13 @@ namespace ferrilata_devilline.IntegrationTests
         public async Task DeleteBadgeApi_CorrectAuthentication_ShouldDeleteLevelsAssociatedWithBadge()
         {
             var request = new HttpRequestMessage(HttpMethod.Delete, "/api/badges/1");
-            request.Headers.Add("Authorization", "test");
+            request.Headers.Add("Authorization", "Bearer " + _tokenService.GenerateToken(email, true));
             var levels = testContext.Context.Badges.GetItemByIndex(0).Levels;
+
             await testContext.Client.SendAsync(request, HttpCompletionOption.ResponseHeadersRead);
-            for (var i = 0; i < levels.Count; i++)
+            for (var i = 0;
+                i < levels.Count;
+                i++)
             {
                 Assert.Empty(testContext.Context.Levels.Where(l => l.LevelId == levels[i].LevelId));
             }
@@ -53,11 +74,14 @@ namespace ferrilata_devilline.IntegrationTests
         public async Task DeleteBadgeApi_CorrectAuthentication_ShouldDeletePitchAssociatedWithBadge()
         {
             var request = new HttpRequestMessage(HttpMethod.Delete, "/api/badges/1");
-            request.Headers.Add("Authorization", "test");
+            request.Headers.Add("Authorization", "Bearer " + _tokenService.GenerateToken(email, true));
             var levels = testContext.Context.Badges.GetItemByIndex(0).Levels;
             var pitches = levels.SelectMany(l => l.Pitches).ToList();
+
             await testContext.Client.SendAsync(request, HttpCompletionOption.ResponseHeadersRead);
-            for (var i = 0; i < pitches.Count; i++)
+            for (var i = 0;
+                i < pitches.Count;
+                i++)
             {
                 Assert.Empty(testContext.Context.Pitches.Where(p => p.PitchId == pitches[i].PitchId));
             }
@@ -67,11 +91,14 @@ namespace ferrilata_devilline.IntegrationTests
         public async Task DeleteBadgeApi_CorrectAuthentication_ShouldDeleteUserLevelAssociatedWithBadge()
         {
             var request = new HttpRequestMessage(HttpMethod.Delete, "/api/badges/1");
-            request.Headers.Add("Authorization", "test");
+            request.Headers.Add("Authorization", "Bearer " + _tokenService.GenerateToken(email, true));
             var levels = testContext.Context.Badges.GetItemByIndex(0).Levels;
             var userLevels = levels.SelectMany(l => l.UserLevels).ToList();
+
             await testContext.Client.SendAsync(request, HttpCompletionOption.ResponseHeadersRead);
-            for (var i = 0; i < userLevels.Count; i++)
+            for (var i = 0;
+                i < userLevels.Count;
+                i++)
             {
                 Assert.Empty(testContext.Context.UserLevels.Where(ul => ul.LevelId == userLevels[i].LevelId));
             }
@@ -81,12 +108,15 @@ namespace ferrilata_devilline.IntegrationTests
         public async Task DeleteBadgeApi_CorrectAuthentication_ShouldDeleteReviewAssociatedWithBadge()
         {
             var request = new HttpRequestMessage(HttpMethod.Delete, "/api/badges/1");
-            request.Headers.Add("Authorization", "test");
+            request.Headers.Add("Authorization", "Bearer " + _tokenService.GenerateToken(email, true));
             var levels = testContext.Context.Badges.GetItemByIndex(0).Levels;
             var pitches = levels.SelectMany(l => l.Pitches).ToList();
             var reviews = pitches.SelectMany(p => p.Reviews).ToList();
+
             await testContext.Client.SendAsync(request, HttpCompletionOption.ResponseHeadersRead);
-            for (var i = 0; i < reviews.Count; i++)
+            for (var i = 0;
+                i < reviews.Count;
+                i++)
             {
                 Assert.Empty(testContext.Context.Reviews.Where(r => r.ReviewId == reviews[i].ReviewId));
             }
@@ -96,9 +126,10 @@ namespace ferrilata_devilline.IntegrationTests
         public async Task DeleteBadgeApi_CorrectAuthentication_ShouldDeletePitchAssociatedWithBadgeFromUser()
         {
             var request = new HttpRequestMessage(HttpMethod.Delete, "/api/badges/1");
-            request.Headers.Add("Authorization", "test");
+            request.Headers.Add("Authorization", "Bearer " + _tokenService.GenerateToken(email, true));
             var levels = testContext.Context.Badges.GetItemByIndex(0).Levels;
             var pitches = levels.SelectMany(l => l.Pitches).ToList();
+
             await testContext.Client.SendAsync(request, HttpCompletionOption.ResponseHeadersRead);
             foreach (var t in pitches)
             {
@@ -111,10 +142,11 @@ namespace ferrilata_devilline.IntegrationTests
         public async Task DeleteBadgeApi_CorrectAuthentication_ShouldDeleteReviewAssociatedWithBadgeFromUser()
         {
             var request = new HttpRequestMessage(HttpMethod.Delete, "/api/badges/1");
-            request.Headers.Add("Authorization", "test");
+            request.Headers.Add("Authorization", "Bearer " + _tokenService.GenerateToken(email, true));
             var levels = testContext.Context.Badges.GetItemByIndex(0).Levels;
             var pitches = levels.SelectMany(l => l.Pitches).ToList();
             var reviews = pitches.SelectMany(p => p.Reviews).ToList();
+
             await testContext.Client.SendAsync(request, HttpCompletionOption.ResponseHeadersRead);
             foreach (var t in reviews)
             {
